+++ packages/ui/src/provider/app-wrapper.tsx
@@ -29,7 +29,7 @@ export function AppWrapper({ children }: Props) {
       <body>
         <QueryClientProviders>
           <Providers>
-            <div className="max-w-3xl mx-auto py-24">
+            <div className="max-w-3xl mx-auto px-4 py-24">
               <Navbar />
               {children}
             </div>