+++ apps/video/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.1.8
+
 ## 0.1.5
 
 ### Patch Changes